Clean build option added in test_with_coverage -script

+++ test/scripts/test_with_coverage.sh
@@ -1,13 +1,31 @@
 #!/bin/bash
 set -e
 
+CLEAN=0
+
+while getopts 'c' OPTION; do
+  case "$OPTION" in
+    c)
+      echo "Cleanup build dir"
+      CLEAN=1
+      ;;
+    ?)
+      echo "script usage: $(basename ${0}) [-c]" >&2
+      exit 1
+      ;;
+  esac
+done
+
 TEST_ROOT_PATH="$(dirname ${0})/.."
 TEST_BUILD_FOLDER="build"
 TEST_BUILD_FOLDER_PATH="${TEST_ROOT_PATH}/${TEST_BUILD_FOLDER}"
 
-# Cleanup
-rm -rf "${TEST_BUILD_FOLDER_PATH}"
-mkdir "${TEST_BUILD_FOLDER_PATH}"
+if [[ ${CLEAN} == 1 ]]
+then
+    rm -rf "${TEST_BUILD_FOLDER_PATH}"
+    mkdir -p "${TEST_BUILD_FOLDER_PATH}"
+fi
+
 cd "${TEST_BUILD_FOLDER_PATH}"
 
 # Build the test application
